+++ src/pages/Students/index.js
@@ -43,19 +43,24 @@ export default function StudentsList() {
   }, [filter, token]);
 
   async function handleDelete(id) {
-    try {
-      await api.delete(`/students/${id}`);
-      toast('Aluno deletado com sucesso', 'success');
+    const result = window.confirm('Desejar deletar esse aluno?');
+    if (result) {
+      try {
+        await api.delete(`/students/${id}`);
 
-      return setTimeout(() => {
-        window.location.reload(false);
-      }, 2000);
-    } catch (err) {
-      if (err.response.data) {
-        return toast(err.response.data.messageContent, 'error');
+        toast('Aluno deletado com sucesso', 'success');
+
+        return setTimeout(() => {
+          window.location.reload(false);
+        }, 2000);
+      } catch (err) {
+        if (err.response.data) {
+          return toast(err.response.data.messageContent, 'error');
+        }
+        return toast('Erro na deleção do aluno.', 'error');
       }
-      return toast('Erro na deleção do aluno.', 'error');
     }
+    return false;
   }
 
   return (
